backend/requirements.txt
fastapi
uvicorn
supabase
pydantic
weasyprint
PyPDF2
python-dotenv
langchain
langchain-openai